change representation of 'Value#' to unlift it

+++ Expr.hs
@@ -16,7 +16,8 @@ import Data.List qualified as List
 import Data.Semigroup ((<>))
 import Data.Text (Text)
 import Data.Text qualified as Text
-import GHC.Base (Double, Type)
+import GHC.Base (Type)
+import GHC.Float (Float)
 import GHC.Show (Show (show))
 
 -- | An expression is a list of atoms.
@@ -36,7 +37,7 @@ data Atom
   | -- | A classic integer literal (most likely to be either 32 or 64-bit).
     AInteger Int
   | -- | A double-precision floating point literal.
-    AFloat Double
+    AFloat Float
   | -- | A unicode character.
     ACharacter Char
   | -- | A classic boolean (true or false).
